feat(app): clean up
