--- src/components/GuideDetailView/GuideDetailHeader/_guide-detail-header.scss
@@ -1,43 +1,0 @@
-@import '../../../styles/responsive';
-@import '@elastic/eui/src/themes/eui-amsterdam/eui_amsterdam_colors_dark.scss';
-
-.guide-header {
-  display: flex;
-  flex: 1;
-  align-items: center;
-  justify-content: space-between;
-  background-color: lighten($euiColorEmptyShade, 2%);
-
-  @include respond-to(gt-sm) {
-    justify-content: flex-end;
-    padding: 8px 0;
-    z-index: 100;
-  }
-
-  @include respond-to(gt-md) {
-    padding: 8px 0;
-    z-index: 100;
-  }
-
-  &__controls__button {
-    margin: 0 4px 0 4px;
-  }
-
-  &__controls {
-    margin-right: 16px;
-  }
-
-  &__title {
-    display: inherit;
-    align-items: inherit;
-    font-size: 24px;
-    font-weight: 600;
-    text-transform: capitalize;
-    margin-left: 12px;
-    margin-right: auto;
-  }
-}
-
-.nav-item {
-  padding: 10px;
-}
