--- src/components/GuideDetailView/GuideDetailSection/_guide-detail-section.scss
@@ -1,121 +1,0 @@
-@import '../../../styles/responsive';
-@import '@elastic/eui/src/themes/eui-amsterdam/eui_amsterdam_colors_dark.scss';
-@import '../../../styles/general';
-
-.guide-section {
-  width: 100%;
-  height: 100%;
-  overflow: auto;
-
-  @include twitter-holder;
-
-  .euiAspectRatio {
-    width: 100%;
-    max-width: 800px;
-    margin-right: auto;
-    margin-left: auto;
-  }
-
-  .euiMarkdownFormat {
-    display: flex;
-    justify-content: center;
-    padding-bottom: 80px !important;
-
-    div {
-      max-width: 900px;
-      width: 100%;
-    }
-
-    table {
-      display: table;
-
-      th {
-        background-color: lighten($color: $euiColorEmptyShade, $amount: 4%);
-      }
-
-      td {
-        background-color: darken($color: $euiColorEmptyShade, $amount: 2%);
-      }
-    }
-  }
-
-  &__title {
-    display: flex;
-    align-items: center;
-    padding: 8px 16px;
-    background-color: lighten($color: $euiColorEmptyShade, $amount: 4%);
-    font-size: 24px;
-    font-weight: 600;
-    text-transform: capitalize;
-    border-bottom: 1px solid $euiColorLightShade;
-    border-top-left-radius: 4px;
-    border-top-right-radius: 4px;
-
-    &--text {
-      margin-right: auto;
-    }
-
-    .euiFormControlLayout {
-      max-width: 200px;
-      margin-right: auto;
-
-      @include respond-to(gt-sm) {
-        max-width: 400px;
-      }
-    }
-  }
-
-  &__body {
-    padding: 16px;
-
-    .euiComboBox .euiComboBox__inputWrap {
-      border-radius: 0;
-    }
-
-    &.editing {
-      height: 100%;
-      padding: 0 !important;
-
-      .euiMarkdownEditorToolbar {
-        border-radius: 0;
-      }
-
-      .euiMarkdownEditorPreview,
-      .euiMarkdownEditorFooter {
-        border-radius: 0;
-      }
-
-      .euiMarkdownEditorFooter {
-        border-radius: 0;
-      }
-    }
-
-    @include respond-to(gt-xs) {
-      padding: 16px;
-    }
-
-    .euiMarkdownFormat {
-      p {
-        display: flex;
-        flex-direction: column;
-        font-size: 18px;
-
-        @include respond-to(gt-sm) {
-          flex-direction: row;
-          flex-wrap: wrap;
-        }
-      }
-
-      .markdown__gif {
-        margin-left: auto;
-        margin-right: auto;
-        width: 100%;
-        max-width: 900px;
-      }
-
-      img {
-        margin: 4px 2px;
-      }
-    }
-  }
-}
