--- src/components/GuideDetailView/GuideDetailSideNav/_guide-detail-side-nav.scss
@@ -1,59 +1,0 @@
-@import '../../../styles/responsive';
-@import '@elastic/eui/src/themes/eui-amsterdam/eui_amsterdam_colors_dark.scss';
-
-.guide-content__side-nav {
-  max-height: 75%;
-  width: 248px;
-  margin-right: 24px;
-  display: none;
-  position: fixed;
-  left: calc(50vw - 700px);
-
-  &.euiPanel {
-    flex-grow: 0;
-
-    &--paddingMedium {
-      padding: 0;
-    }
-  }
-
-  @include respond-to(gt-sm) {
-    margin-left: 0;
-    margin-right: 12px;
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    text-transform: capitalize;
-  }
-
-  .side-nav {
-    &__header {
-      display: flex;
-      align-items: center;
-      font-size: 24px;
-      font-weight: 600;
-      text-transform: capitalize;
-      width: 100%;
-      height: 57px;
-    }
-
-    &__content {
-      display: flex;
-      flex-direction: column;
-      align-items: center;
-      padding: 16px;
-      width: 100%;
-      &__items {
-        margin: 12px 0;
-      }
-    }
-  }
-  &__item {
-    display: flex;
-    align-items: center;
-  }
-
-  &__handle {
-    margin-right: 10px;
-  }
-}
