--- src/components/GuideListView/GuideCard/_guide-card.scss
@@ -1,83 +1,0 @@
-@import '../../../styles/responsive';
-@import '@elastic/eui/src/themes/eui-amsterdam/eui_amsterdam_colors_dark.scss';
-
-.guide-card {
-  min-width: 300px;
-  min-height: 150px;
-  max-height: 150px;
-  cursor: pointer;
-
-  // @include respond-to(gt-xs) {
-  //   max-width: 400px;
-  // }
-
-  &.euiPanel {
-    flex-grow: 0;
-
-    &--paddingMedium {
-      padding: 0;
-    }
-
-    &.euiPanel--isClickable:enabled {
-      display: flex;
-      flex-direction: column;
-      align-items: center;
-      text-transform: capitalize;
-    }
-  }
-
-  &__header {
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
-    padding: 8px 16px;
-    background-color: lighten($color: $euiColorEmptyShade, $amount: 4%);
-    font-size: 24px;
-    font-weight: 600;
-    text-transform: capitalize;
-    width: 100%;
-    height: 57px;
-    border-top-left-radius: 4px;
-    border-top-right-radius: 4px;
-
-    &__title,
-    &__controls {
-      display: inherit;
-      align-items: inherit;
-    }
-
-    &__delete {
-      margin-left: 10px;
-    }
-
-    &__avatar {
-      margin-right: 14px;
-    }
-  }
-
-  &__content {
-    position: relative;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    width: 100%;
-    height: 100%;
-
-    &__description {
-      text-transform: none;
-      margin-bottom: 32px;
-      margin-top: 12px;
-      font-weight: 500;
-      font-size: 14px;
-    }
-
-    &__footer {
-      position: absolute;
-      left: 6px;
-      bottom: 6px;
-    }
-    .euiBadge {
-      text-transform: lowercase;
-    }
-  }
-}
