--- src/components/GuideListView/_guide-list-view.scss
@@ -1,30 +1,0 @@
-@import '../../styles/responsive';
-
-.guide-list {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  overflow: hidden;
-  width: 100%;
-  max-width: 800px;
-  padding: 0;
-
-  &__content {
-    display: grid;
-    grid-template-columns: repeat(auto-fill, 100%);
-    grid-gap: 10px;
-    width: calc(100% - 16px);
-    justify-content: center;
-    overflow: auto;
-    padding-top: 8px;
-    padding-bottom: 48px;
-
-    @include respond-to(gt-xs) {
-      width: 100%;
-    }
-
-    .euiFlexItem {
-      margin-bottom: 0 !important;
-    }
-  }
-}
