+++ src/components/PostListView/PostView/_post-view.scss
@@ -45,7 +45,6 @@
 
   .euiAspectRatio {
     width: 100%;
-    max-width: 800px;
     margin-right: auto;
     margin-left: auto;
   }
