+++ src/components/PostListView/_post-list-view.scss
@@ -26,7 +26,6 @@
 
   &__content {
     width: 100%;
-    max-width: 900px;
     height: 100%;
     overflow: auto;
     display: flex;
@@ -34,6 +33,8 @@
     align-items: center;
     padding-bottom: 48px;
     padding-top: 8px;
+    padding-left: 8px;
+    padding-right: 8px;
 
     @include respond-to(gt-md) {
       overflow: visible;
